+++ lib/struct_fields.ex
@@ -1,32 +1,46 @@
 defmodule StructFields do
   @moduledoc """
-  Defines a `fields` macro that returns all fields of a struct.
+  Defines a `fields` function that returns all fields of a struct.
+
+  Fields can be ignored by using the `ignore` option; the `:__struct__` field is
+  always ignored.
 
   ## Examples
 
-      defmodule MyModule do
-        defstruct [:foo, :bar]
+      iex> defmodule MyModule do
+      ...>   use StructFields
+      ...>
+      ...>   defstruct [:foo, :bar]
+      ...> end
+      ...>
+      ...> MyModule.fields
+      [:bar, :foo]
 
-        # Use the module
-        use StructFields
-      end
-
-      MyModule.fields
-      #=> [:foo, :bar]
+      iex> defmodule AnotherModule do
+      ...>  use StructFields, ignore: [:not_me]
+      ...>
+      ...>  defstruct [:this, :not_me, :that]
+      ...> end
+      ...>
+      ...> AnotherModule.fields
+      [:that, :this]
   """
 
   @doc false
-  defmacro __using__(_opts) do
-    quote do
+  defmacro __using__(opts) do
+    fields_to_ignore = [:__struct__ | Dict.get(opts, :ignore, [])]
+    quote bind_quoted: [fields_to_ignore: fields_to_ignore] do
+      Module.put_attribute __MODULE__, :fields_to_ignore, fields_to_ignore
+
       @before_compile StructFields
     end
   end
 
   defmacro __before_compile__(_env) do
     quote do
-      def fields do
-        Map.keys(%__MODULE__{}) |> List.delete(:__struct__)
-      end
+      @fields Map.keys(%__MODULE__{}) -- @fields_to_ignore
+
+      def fields, do: @fields
     end
   end
 end