+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+- Add `ignore` option to exclude fields
+
 ## [0.2.0] - 2016-02-07
 - Allow `use StructFields` to be placed above struct definition